requirements.txt
fastapi
uvicorn[standard]
gunicorn
requests
beautifulsoup4
nltk
pydantic
python-dotenv
aioredis
